refactor(core): prepare semver boundary for extraction
