+++ .github/workflows/ci-test.yml
@@ -20,6 +20,6 @@ jobs:
     - name: Check compilation
       run: cargo check
     - name: Lint
-      run: cargo clippy --all-targets --all-features
+      run: cargo clippy --all-targets --all-features -- -D warnings
     - name: Run tests
       run: cargo test
