+++ scripts/install-git-hooks.sh
@@ -18,4 +18,5 @@ git config --local core.hooksPath .githooks
 
 echo "Installed repo-local git hooks at .githooks"
 echo "pre-commit: cargo fmt --check"
+echo "pre-push: cargo clippy --all-targets --all-features -- -D warnings"
 echo "pre-push: cargo test"
