+++ src/core/resolver/semver/range/mod.rs
@@ -1,0 +1,10 @@
+mod construct;
+pub(crate) mod desugar;
+pub(crate) mod evaluate;
+pub(crate) mod interval;
+pub(crate) mod normalize;
+pub(crate) mod parse;
+mod types;
+
+pub use types::Range;
+pub(crate) use types::{Comparator, ComparatorOp, ComparatorSet};
