+++ src/core/resolver/semver/tests/mod.rs
@@ -1,0 +1,1 @@
+mod compatibility;
